Move FF synchronizer to its own package
