+++ app/featureflag/split_logger.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-package app
+package featureflag
 
 import (
 	"fmt"
